+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 11.0.2-20240910-1
 
 * Complete refactor and code cleanup.
+* Added The War Within factions to faction filter.
 
 ## v0.43.3